# Enable inline requires