--- a/babel.config.js
+++ b/babel.config.js
@@ -1,7 +1,14 @@
 module.exports = function (api) {
   api.cache(true)
   return {
-    presets: ['babel-preset-expo'],
+    presets: [
+      [
+        'babel-preset-expo',
+        {
+          lazyImports: true,
+        },
+      ],
+    ],
     plugins: [
       [
         'module:react-native-dotenv',